requirements.txt
openai
langchain
python-dotenv
Flask
openai
pyyaml